requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.29.3
colorama==0.4.6
Django==5.1
djangorestframework==3.15.2
exceptiongroup==1.1.1
filelock==3.15.4
Flask==2.2.3
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==3.3.3
Levenshtein==0.25.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
outcome==1.2.0
packaging==23.0
Pillow==9.5.0
protobuf==4.25.3
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==1.0.0
pytweening==1.0.4
pywhatkit==5.4
rapidfuzz==3.9.4
requests==2.28.2
requests-file==2.1.0
rich==13.7.1
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.5.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tldextract==5.1.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.12.1
tzdata==2024.1
urllib3==1.26.15
webdriver-manager==3.8.5
Werkzeug==2.2.3
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
